+++ src/lib/pack/import/import.sh
@@ -17,9 +17,6 @@
 source /usr/local/lib/doil/lib/include/env.sh
 source /usr/local/lib/doil/lib/include/helper.sh
 
-# we need to move the pointer two positions
-shift
-
 # check if command is just plain help
 # if we don't have any command we load the help
 POSITION=1
@@ -57,7 +54,7 @@ done
 
 if [[ -z "${INSTANCE}" ]]
 then
-  read -p "Name the instance you'd like to import: " INSTANCE
+  read -p "Name the instance you'd like to import to: " INSTANCE
 fi
 if [[ ${GLOBAL} == TRUE ]]
 then
@@ -153,40 +150,46 @@ doil_send_log "Copying necessary files"
 doil down ${INSTANCE} ${FLAG} --quiet
 
 # remove all the files
-sudo rm -rf ${TARGET}/volumes/data
-sudo rm -rf ${TARGET}/volumes/ilias/data
-sudo rm -rf ${TARGET}/volumes/ilias/ilias.ini.php
+rm -rf ${TARGET}/volumes/data
+rm -rf ${TARGET}/volumes/ilias/data
+rm -rf ${TARGET}/volumes/ilias/ilias.ini.php
 
 # import the files
-sudo mkdir -p ${TARGET}/volumes/ilias/data/
-sudo mkdir -p ${TARGET}/volumes/data/
-sudo cp -r ${PWD}/${PACKNAME}/var/www/html/ilias.ini.php ${TARGET}/volumes/ilias/ilias.ini.php
-sudo cp -r ${PWD}/${PACKNAME}/var/www/html/data ${TARGET}/volumes/ilias/
-sudo cp -r ${PWD}/${PACKNAME}/var/ilias/data/* ${TARGET}/volumes/data
-sudo cp -r ${PWD}/${PACKNAME}/var/ilias/ilias.sql ${TARGET}/volumes/data/ilias.sql
-
-sudo chown -R ${USER}:${USER} ${TARGET}
+mkdir -p ${TARGET}/volumes/ilias/data/
+mkdir -p ${TARGET}/volumes/data/
+cp -r ${PWD}/${PACKNAME}/var/www/html/ilias.ini.php ${TARGET}/volumes/ilias/ilias.ini.php
+cp -r ${PWD}/${PACKNAME}/var/www/html/data ${TARGET}/volumes/ilias/
+cp -r ${PWD}/${PACKNAME}/var/ilias/data/* ${TARGET}/volumes/data
+cp -r ${PWD}/${PACKNAME}/var/ilias/ilias.sql ${TARGET}/volumes/data/ilias.sql
 
 # start the instance
 doil up ${INSTANCE} --quiet ${FLAG}
-docker exec -ti ${INSTANCE}_${SUFFIX} bash -c "chown -R mysql:mysql /var/lib/mysql"
-docker exec -ti ${INSTANCE}_${SUFFIX} bash -c "service mysql restart"
-sleep 5
+sleep 15
 doil_send_log "Importing database"
 
 # import database
-echo "Please enter your MySQL password: "
-read -s SQLPW
-echo "Ok"
+if [[ -f "${TARGET}/README.md" ]]
+then
+  SQLPW=$(cat "${TARGET}/README.md" | grep MYSQL_PASSWORD | cut -d\   -f2)
+else
+  echo "Please enter your MySQL password: "
+  read -s SQLPW
+  echo "Ok"
+fi
 
-touch ${TARGET}/volumes/data/mysql-client.conf
-echo "[client]" > ${TARGET}/volumes/data/mysql-client.conf
-echo "user=ilias" >> ${TARGET}/volumes/data/mysql-client.conf
-echo "password=${SQLPW}"  >> ${TARGET}/volumes/data/mysql-client.conf
+#touch ${TARGET}/volumes/data/mysql-client.conf
+#echo "[client]" > ${TARGET}/volumes/data/mysql-client.conf
+#echo "user=ilias" >> ${TARGET}/volumes/data/mysql-client.conf
+#echo "password=${SQLPW}" >> ${TARGET}/volumes/data/mysql-client.conf
 
-docker exec -ti ${INSTANCE}_${SUFFIX} bash -c 'mysql --defaults-extra-file=/var/ilias/data/mysql-client.conf -e "DROP DATABASE IF EXISTS ilias;"'
-docker exec -ti ${INSTANCE}_${SUFFIX} bash -c 'mysql --defaults-extra-file=/var/ilias/data/mysql-client.conf -e "CREATE DATABASE ilias;"'
-docker exec -ti ${INSTANCE}_${SUFFIX} bash -c "mysql --defaults-extra-file=/var/ilias/data/mysql-client.conf ilias < /var/ilias/data/ilias.sql"
+#docker exec -i ${INSTANCE}_${SUFFIX} bash -c '/etc/init.d/mariadb stop'
+#docker exec -i ${INSTANCE}_${SUFFIX} bash -c 'chown -R mysql:mysql /var/lib/mysql'
+#docker exec -i ${INSTANCE}_${SUFFIX} bash -c 'chown -R root:root /etc/mysql'
+#docker exec -i ${INSTANCE}_${SUFFIX} bash -c "service mysql restart"
+
+docker exec -i ${INSTANCE}_${SUFFIX} bash -c 'mysql -e "DROP DATABASE IF EXISTS ilias;"'
+docker exec -i ${INSTANCE}_${SUFFIX} bash -c 'mysql -e "CREATE DATABASE ilias;"'
+docker exec -i ${INSTANCE}_${SUFFIX} bash -c "mysql ilias < /var/ilias/data/ilias.sql"
 
 CLIENT_FILE_LOCATION=$(find ${TARGET}/volumes/ilias/data/ -iname client.ini.php)
 sed -i "s/pass =.*/pass = '${SQLPW}'/" ${CLIENT_FILE_LOCATION}
@@ -194,9 +197,9 @@ sed -i "s/pass =.*/pass = '${SQLPW}'/" ${CLIENT_FILE_LOCATION}
 doil_send_log "Setting permissions"
 
 # set access
-sudo chown -R ${USER}:${USER} ${TARGET}
-docker exec -ti ${INSTANCE}_${SUFFIX} bash -c "chown -R mysql:mysql /var/lib/mysql"
-docker exec -ti ${INSTANCE}_${SUFFIX} bash -c "service mysql restart"
+#sudo chown -R ${USER}:${USER} ${TARGET}
+#docker exec -i ${INSTANCE}_${SUFFIX} bash -c "chown -R mysql:mysql /var/lib/mysql"
+#docker exec -i ${INSTANCE}_${SUFFIX} bash -c "service mysql restart"
 
 doil down ${INSTANCE} --quiet ${FLAG}
 doil up ${INSTANCE} --quiet ${FLAG}
@@ -206,6 +209,6 @@ doil apply ${INSTANCE} access --quiet ${FLAG}
 doil_send_log "Cleanup"
 
 rm -rf ${PWD}/${PACKNAME}
-docker exec -ti ${INSTANCE}_${SUFFIX} bash -c "rm /var/ilias/data/ilias.sql"
+docker exec -i ${INSTANCE}_${SUFFIX} bash -c "rm /var/ilias/data/ilias.sql"
 
 doil_send_log "Import of ${INSTANCE} done"