+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## 1.4.4
+
+* Bulletproofed windows integration
+* implemented better volume handling
+* increased speed for `doil create` by using predefined container
+* fixed several salt stack issues
+* implemented `-y` flag for `doil delete`
+* fixed key check
+* updated password generation
+* removed `-t` flag for docker exec functions
+* fixed cron service start after `doil up`
+* updated ilias config with webservices
+
+## 1.4.3
+
+* Fixed port problem in minion services
+
+## 1.4.2
+
+* Added compatibility to Windows and WSL with ubuntu
+* Implemented improvements to `doil pack`
+
+## 1.4.1
+
+* Fixed small issues with permissions
+
 ## 1.4
 
 * Implemented globla user support